+++ test/benchmarking/test_timer.cpp
@@ -1,0 +1,73 @@
+#include <thread>
+
+#include "gtest/gtest.h"
+#include "perf/benchmarking.hpp"
+
+void waitfor_function(const size_t ms) {
+  std::this_thread::sleep_for(std::chrono::milliseconds(ms));
+}
+
+// chrono
+TEST(timer, is_elapsed_time_returns_positive_value) {
+  const size_t a = 250;
+  double res_time;
+  res_time = elapsed_time<double, std::milli>(waitfor_function, a);
+  EXPECT_GE(res_time, 0.0);
+}
+TEST(timer, is_elapsed_time_avg_returns_positive_value) {
+  const size_t a = 250;
+  const size_t b = 10;
+  double res_time;
+  res_time = elapsed_time_avg<double, std::milli>(b, waitfor_function, a);
+  EXPECT_GE(res_time, 0.0);
+}
+
+// omp
+TEST(timer, is_elapsed_time_omp_returns_positive_value) {
+  const size_t a = 250;
+  double res_time;
+  res_time = elapsed_time_omp(waitfor_function, a);
+  EXPECT_GE(res_time, 0.0);
+}
+TEST(timer, is_elapsed_time_omp_avg_returns_positive_value) {
+  const size_t a = 250;
+  const size_t b = 10;
+  double res_time;
+  res_time = elapsed_time_omp_avg(b, waitfor_function, a);
+  EXPECT_GE(res_time, 0.0);
+}
+
+// >= -100 ms, <= +1000 ms
+// chrono
+TEST(timer, is_elapsed_time_returns_nearly_correct_time) {
+  const size_t a = 250;
+  double res_time;
+  res_time = elapsed_time<double, std::milli>(waitfor_function, a);
+  EXPECT_GE(res_time, 150);
+  EXPECT_LE(res_time, 1250);
+}
+TEST(timer, is_elapsed_time_avg_returns_nearly_correct_time) {
+  const size_t a = 250;
+  const size_t b = 10;
+  double res_time;
+  res_time = elapsed_time_avg<double, std::milli>(b, waitfor_function, a);
+  EXPECT_GE(res_time, 150);
+  EXPECT_LE(res_time, 1250);
+}
+
+// omp
+TEST(timer, is_elapsed_time_omp_returns_nearly_correct_time) {
+  const size_t a = 250;
+  double res_time;
+  res_time = elapsed_time_omp(waitfor_function, a);
+  EXPECT_GE(res_time, 0.15);
+  EXPECT_LE(res_time, 1.25);
+}
+TEST(timer, is_elapsed_time_omp_avg_returns_nearly_correct_time) {
+  const size_t a = 250;
+  const size_t b = 10;
+  double res_time;
+  res_time = elapsed_time_omp_avg(b, waitfor_function, a);
+  EXPECT_GE(res_time, 0.15);
+  EXPECT_LE(res_time, 1.25);
+}
